+++ WPF-SWL_Dice_Calculator/LIB_SWL_Dice_Calculator/ResultModels/DiceResult.cs
@@ -8,7 +8,7 @@ namespace LIB_SWL_Dice_Calculator.ResultModels
 {
     public abstract class DiceResult
     {
-        public int Blanks { get; set; }
-        public int Surges { get; set; }
+        public float Blanks { get; set; }
+        public float Surges { get; set; }
     }
 }
